fix hwapi package for non x86 system
