+++ pkg/hwapi/uint.go
@@ -2,19 +2,12 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// +build !arm,!arm64
-
 package hwapi
 
 import (
-	"encoding/binary"
 	"fmt"
-	"io"
-	"os"
 	"unsafe"
 )
-
-var memPaths = [...]string{"/dev/fmem", "/dev/mem"}
 
 // UintN is a wrapper around uint types and provides a few io-related
 // functions.
@@ -124,86 +117,4 @@ func (u *Uint64) write(addr unsafe.Pointer) error {
 	// Warning: On arm, this uses two str's rather than strd.
 	*(*uint64)(addr) = uint64(*u) // TODO: rewrite in Go assembly for ARM
 	return nil                    // TODO: catch misalign, segfault, sigbus, ...
-}
-
-func pathRead(path string, addr int64, data UintN) error {
-	f, err := os.OpenFile(path, os.O_RDONLY, 0)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-
-	if _, err := f.Seek(addr, io.SeekCurrent); err != nil {
-		return err
-	}
-	return binary.Read(f, binary.LittleEndian, data)
-}
-
-func selectDevMem() (string, error) {
-	if len(memPaths) == 0 {
-		return "", fmt.Errorf("Internal error: no /dev/mem device specified")
-	}
-
-	for _, p := range memPaths {
-		if _, err := os.Stat(p); err == nil {
-			return p, nil
-		}
-	}
-
-	return "", fmt.Errorf("No suitable /dev/mem device found. Tried %#v", memPaths)
-}
-
-// ReadPhys reads data from physical memory at address addr. On x86 platforms,
-// this uses the seek+read syscalls.
-func (t TxtAPI) ReadPhys(addr int64, data UintN) error {
-	devMem, err := selectDevMem()
-	if err != nil {
-		return err
-	}
-
-	return pathRead(devMem, addr, data)
-}
-
-// ReadPhysBuf reads data from physical memory at address addr. On x86 platforms,
-// this uses the seek+read syscalls.
-func (t TxtAPI) ReadPhysBuf(addr int64, buf []byte) error {
-	devMem, err := selectDevMem()
-	if err != nil {
-		return err
-	}
-
-	f, err := os.OpenFile(devMem, os.O_RDONLY, 0)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-
-	if _, err := f.Seek(addr, io.SeekCurrent); err != nil {
-		return err
-	}
-	return binary.Read(f, binary.LittleEndian, buf)
-}
-
-func pathWrite(path string, addr int64, data UintN) error {
-	f, err := os.OpenFile(path, os.O_WRONLY, 0)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-
-	if _, err := f.Seek(addr, io.SeekCurrent); err != nil {
-		return err
-	}
-	return binary.Write(f, binary.LittleEndian, data)
-}
-
-// WritePhys writes data to physical memory at address addr. On x86 platforms, this
-// uses the seek+read syscalls.
-func (t TxtAPI) WritePhys(addr int64, data UintN) error {
-	devMem, err := selectDevMem()
-	if err != nil {
-		return err
-	}
-
-	return pathWrite(devMem, addr, data)
 }
